move shared schemas into the separate folder
