+++ src/schemas/faqWriteQuestionValidationSchema.ts
@@ -1,31 +1,17 @@
-import { useTranslations } from "next-intl";
 import * as yup from "yup";
 
-import {
-    emailRegex,
-    nameRegex,
-    wrongRuEmailsRegex,
-} from "./writeUsFormValidationSchema";
+import { EmailValidation } from "./shared/emailValidation";
+import { MessageValidation } from "./shared/messageValidation";
+import { NameValidation } from "./shared/nameValidation";
 
 export const FaqWriteQuestionValidation = () => {
-    const getTranslation = useTranslations("Errors");
+    const nameValidation = NameValidation();
+    const emailValidation = EmailValidation();
+    const messageValidation = MessageValidation();
+
     return yup.object().shape({
-        name: yup
-            .string()
-            .min(2, getTranslation("nameMinMaxSymbols"))
-            .max(30, getTranslation("nameMinMaxSymbols"))
-            .matches(nameRegex, getTranslation("nameAllowedSymbols"))
-            .required(getTranslation("required")),
-        email: yup
-            .string()
-            .matches(emailRegex, getTranslation("wrongEmail"))
-            .matches(wrongRuEmailsRegex, getTranslation("russianEmails"))
-            .required(getTranslation("required")),
-        message: yup
-            .string()
-            .trim()
-            .min(10, getTranslation("messageMinMaxSymbols"))
-            .max(300, getTranslation("messageMinMaxSymbols"))
-            .required(getTranslation("required")),
+        name: nameValidation,
+        email: emailValidation,
+        message: messageValidation,
     });
 };
