+++ src/schemas/shared/messageValidation.ts
@@ -1,0 +1,13 @@
+import { useTranslations } from "next-intl";
+import * as yup from "yup";
+
+export const MessageValidation = () => {
+    const getTranslation = useTranslations("Errors");
+
+    return yup
+        .string()
+        .trim()
+        .min(10, getTranslation("messageMinMaxSymbols"))
+        .max(300, getTranslation("messageMinMaxSymbols"))
+        .required(getTranslation("required"));
+};
