+++ src/schemas/shared/nameValidation.ts
@@ -1,0 +1,15 @@
+import { useTranslations } from "next-intl";
+import * as yup from "yup";
+
+const nameRegex = /^[a-zA-ZąćęłńóśźżĄĆĘŁŃÓŚŹŻіІїЇґҐєЄа-яА-Я'"`\s-]+$/;
+
+export const NameValidation = () => {
+    const getTranslation = useTranslations("Errors");
+
+    return yup
+        .string()
+        .min(2, getTranslation("nameMinMaxSymbols"))
+        .max(30, getTranslation("nameMinMaxSymbols"))
+        .matches(nameRegex, getTranslation("nameAllowedSymbols"))
+        .required(getTranslation("required"));
+};
